backend/requirements.txt
flask>=3.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.1.0
beautifulsoup4>=4.12.0
